chore: renaming and project cleanup
