+++ backend/getGroupedImages.test.ts
