+++ backend/getGroupedImages.ts
