+++ backend/updateBaseImagesInS3.ts
