+++ frontend/components/ImageView.tsx
