+++ frontend/components/UpdateImagesButton.tsx
