+++ frontend/providers/BaseImageStateProvider.tsx
